+++ cmd/frontend/graphqlbackend/schema.graphql
@@ -1988,8 +1988,6 @@ type User implements Node & ConfigurationSubject {
     # Only the user and site admins can access this field.
     tags: [String!]!
     # The user's usage statistics on Sourcegraph.
-    #
-    # Only the user and site admins can access this field.
     usageStatistics: UserUsageStatistics!
     # The user's email addresses.
     #
@@ -2660,6 +2658,8 @@ type Configuration {
 }
 
 # UserUsageStatistics describes a user's usage statistics.
+#
+# This information is visible to all viewers.
 type UserUsageStatistics {
     # The number of search queries that the user has performed.
     searchQueries: Int!
@@ -2694,6 +2694,8 @@ enum UserActivePeriod {
 }
 
 # SiteUsageStatistics describes a site's aggregate usage statistics.
+#
+# This information is visible to all viewers.
 type SiteUsageStatistics {
     # Recent daily active users.
     daus: [SiteUsagePeriod!]!
@@ -2704,6 +2706,8 @@ type SiteUsageStatistics {
 }
 
 # SiteUsagePeriod describes a site's usage statistics for a given timespan.
+#
+# This information is visible to all viewers.
 type SiteUsagePeriod {
     # The time when this started.
     startTime: String!
