+++ cmd/frontend/graphqlbackend/site_usage_stats.go
@@ -1,10 +1,42 @@
 package graphqlbackend
 
 import (
+	"context"
+	"errors"
 	"time"
 
+	"github.com/sourcegraph/sourcegraph/cmd/frontend/envvar"
+	"github.com/sourcegraph/sourcegraph/cmd/frontend/internal/pkg/usagestats"
 	"github.com/sourcegraph/sourcegraph/cmd/frontend/types"
 )
+
+func (r *siteResolver) UsageStatistics(ctx context.Context, args *struct {
+	Days   *int32
+	Weeks  *int32
+	Months *int32
+}) (*siteUsageStatisticsResolver, error) {
+	if envvar.SourcegraphDotComMode() {
+		return nil, errors.New("site usage statistics are not available on sourcegraph.com")
+	}
+	opt := &usagestats.SiteUsageStatisticsOptions{}
+	if args.Days != nil {
+		d := int(*args.Days)
+		opt.DayPeriods = &d
+	}
+	if args.Weeks != nil {
+		w := int(*args.Weeks)
+		opt.WeekPeriods = &w
+	}
+	if args.Months != nil {
+		m := int(*args.Months)
+		opt.MonthPeriods = &m
+	}
+	activity, err := usagestats.GetSiteUsageStatistics(opt)
+	if err != nil {
+		return nil, err
+	}
+	return &siteUsageStatisticsResolver{activity}, nil
+}
 
 type siteUsageStatisticsResolver struct {
 	siteUsageStatistics *types.SiteUsageStatistics
