+++ cmd/frontend/graphqlbackend/user_usage_stats_test.go
@@ -1,17 +1,48 @@
 package graphqlbackend
 
 import (
-	"context"
 	"testing"
 
+	"github.com/graph-gophers/graphql-go/gqltesting"
+	"github.com/sourcegraph/sourcegraph/cmd/frontend/db"
+	"github.com/sourcegraph/sourcegraph/cmd/frontend/internal/pkg/usagestats"
 	"github.com/sourcegraph/sourcegraph/cmd/frontend/types"
 )
 
-func TestUsers_UsageStatistics(t *testing.T) {
-	ctx := context.Background()
-	u := &UserResolver{user: &types.User{}}
-	_, err := u.UsageStatistics(ctx)
-	if err == nil {
-		t.Errorf("Non-admin, non-user can access endpoint")
+func TestUser_UsageStatistics(t *testing.T) {
+	resetMocks()
+	db.Mocks.Users.MockGetByID_Return(t, &types.User{ID: 1, Username: "alice"}, nil)
+	usagestats.MockGetByUserID = func(userID int32) (*types.UserUsageStatistics, error) {
+		return &types.UserUsageStatistics{
+			SearchQueries: 2,
+		}, nil
 	}
+	defer func() { usagestats.MockGetByUserID = nil }()
+	gqltesting.RunTests(t, []*gqltesting.Test{
+		{
+			Schema: GraphQLSchema,
+			Query: `
+				{
+					node(id: "VXNlcjox") {
+						id
+						... on User {
+							usageStatistics {
+								searchQueries
+							}
+						}
+					}
+				}
+			`,
+			ExpectedResult: `
+				{
+					"node": {
+						"id": "VXNlcjox",
+						"usageStatistics": {
+							"searchQueries": 2
+						}
+					}
+				}
+			`,
+		},
+	})
 }
