+++ doc/user/usage_statistics.md
@@ -1,7 +1,11 @@
-# Usage analytics
+# Usage statistics
 
-Sourcegraph records basic usage statistics for site admins. To view analytics, visit the **Site admin > Analytics** page. (The URL is `https://sourcegraph.example.com/site-admin/analytics`.)
+Sourcegraph records basic per-user usage statistics. To view analytics, visit the **Site admin > Usage stats** page. (The URL is `https://sourcegraph.example.com/site-admin/usage-statistics`.) This information is available via the GraphQL API to all viewers (not just site admins).
 
 Here you can see charts with counts of unique users by day, week, or month, split into authenticated and anonymous users.
 
 From this page, you can also see user-level activity, including counts of pageviews, searches, and code intelligence actions, and last active times. This user-level data is all stored locally on your Sourcegraph instance, and is never sent to Sourcegraph.com.
+
+## See also 
+
+- [User satisfaction surveys](user_surveys.md)
