--- doc/user/usage_statistics/index.md
@@ -1,8 +1,0 @@
-# Usage statistics
-
-Sourcegraph collects 2 kinds of usage statistics to track engagement and satisfaction:
-
-- [Usage analytics](analytics.md) (daily/weekly/monthly active users)
-- [User satisfaction surveys](user_surveys.md)
-
-The results are shown only to site admins.
