+++ doc/user/user_surveys.md
@@ -1,4 +1,4 @@
-## User surveys
+# User surveys
 
 After using Sourcegraph for a few days, users are presented with a request to answer "How likely is it that you would recommend Sourcegraph to a friend?", and to provide some qualitative product feedback.
 
@@ -7,3 +7,7 @@ Responses to this survey are visible to all site admins on the Sourcegraph insta
 On this page, admins can view individual satisfaction scores (from 0–10) and individual qualitative feedback, along with aggregate historical scores.
 
 Survey responses are also always sent to Sourcegraph.com.
+
+## See also 
+
+- [Usage statistics](usage_statistics.md)
